chore: adjust web component styles (#362)
* chore: adjust web component styles

* docs: update notes
