+++ docs/content.en/docs/release-notes/_index.md
@@ -30,6 +30,7 @@ Information about release notes of Coco Server is provided here.
 
 - refactor: web components #331
 - refactor: refactoring login callback, receive access_token from coco-server
+- chore: adjust web component styles #362
 
 ## 0.3.0 (2025-03-31)
